fix NPE in MultiSegmentIterator when iterator starts after last index change parameter for AppendOnlyFileStore.iterator(fromIndex) from int to long

+++ axonserver-filestore/src/main/java/io/axoniq/axonserver/filestorage/AppendOnlyFileStore.java
@@ -67,7 +67,7 @@ public interface AppendOnlyFileStore {
      * @return iterator of entries
      */
     // TODO: 12/30/21 remove
-    CloseableIterator<FileStoreEntry> iterator(int fromIndex);
+    CloseableIterator<FileStoreEntry> iterator(long fromIndex);
 
     /**
      * Deletes the file store.
